split out parser into yargs-parser library, and make yargs configurable see #352
