+++ package.json
@@ -18,7 +18,8 @@
     "read-pkg-up": "^1.0.1",
     "string-width": "^1.0.1",
     "window-size": "^0.1.4",
-    "y18n": "^3.2.0"
+    "y18n": "^3.2.0",
+    "yargs-parser": "^1.1.0"
   },
   "devDependencies": {
     "chai": "^3.4.1",
@@ -101,5 +102,8 @@
   "license": "MIT",
   "engine": {
     "node": ">=0.10"
+  },
+  "yargs": {
+    "camel-case-expansion": true
   }
 }
